simplify list output
For lists, plain old indented paragraphs are the perfect tool.  To not
have vertical space between list items, we use .PD 0.

Should the many needless .PP be removed, will the .PD 0 have to be moved
after the first list item, such that that introduces vertical space
after the last paragraph.